Create useHeaderCount hook to load user location count before global count
